Fix end of buffer race condition
In the bootloader we have a "buffer_send_pointer == buffer_send_pointer_end"
check to find out if a response is fully send.

If the fifo is one byte smaller then the amount of data to be send
this check can become a false positive. To fix this we check for this
case in the interrupt and wait for one byte to free up