Improve SPI IRQ performance and disable other IRQs during transfer
Use local pointer to save the time for accessing the struct.

Addisionally disable all other IRQs while we write the data. If we don't
disable the IRQs here, there is a minor race condition:
If a IRQ with higher priority is triggered exactly after the first byte is
written to the FIFO and this IRQ takes longer then it takes to transfer one
byte we will send erronous data (instead of A B C we send A 0 B C).
The CRC of this message does not fit, so this will be resolved through the
Brick by requesting the message again.

However, for Bricklets that do not hafe any real-time requirement for
the high priority IRQs (or nor IRQs), we can disable the irqs here.

Bricklets that have real-time requirement for IRQs have to set the
"SPITFP_NOT_ALLOWED_TO_DISABLE_IRQ" define.