warp/sdm: Reset timeout counter when meter is found for the first time.
There is an inherent race condition between the start-up of the EVSE
or Energy Manager Bricklet and the energy meter. If the bricklet is
faster, attempting to read the meter type creates timeouts.

Reset the (local) timeout counter if reading the meter type succeeds.
The meter type is read only once (for supported meters).